# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SmartRoute API - Delivery route optimization
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: smartroute_token
security:
  - cookieAuth: []
paths:
...
  /route/build:
    post:
      operationId: buildRoute
      tags: [routes]
      summary: Build optimized delivery routes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Optimized routes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResult"
  /route/sessions:
    get:
      operationId: listRouteSessions
      tags: [routes]
      summary: List route sessions (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of route sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteSessionList"
  /route/sessions/{id}:
    get:
      operationId: getRouteSession
      tags: [routes]
      summary: Get a saved route session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full route session data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResult"
        "404":
          description: Not found
    delete:
      operationId: deleteRouteSession
      tags: [routes]
      summary: Delete a route session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "404":
          description: Not found
...
components:
  schemas:
...
    RouteRequest:
      type: object
      properties:
        store_ids:
          type: array
          items:
            type: integer
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/VehicleInput"
        depot_lat:
          type: ["number", "null"]
        depot_lon:
          type: ["number", "null"]
        use_time_windows:
          type: boolean
        use_unload_time:
          type: boolean
        max_stops_per_vehicle:
          type: ["integer", "null"]
          description: "Мягкий лимит точек на машину. Рекомендуется 24 при дисбалансе (ratio >3x). Параметр опциональный — null = без ограничения."
      required:
        - store_ids
        - vehicles
    VehicleInput:
      type: object
      properties:
        name:
          type: string
        capacity_kg:
          type: ["integer", "null"]
        capacity_m3:
          type: ["number", "null"]
          description: "Объём кузова (м³). Независимое ограничение от kg. null = без ограничения по объёму."
        average_speed:
          type: ["number", "null"]
          description: Average speed in km/h for this vehicle (overrides global default)
      required:
        - name
    RouteResult:
      type: object
      properties:
        session_id:
          type: ["integer", "null"]
        routes:
          type: array
          items:
            $ref: "#/components/schemas/VehicleRoute"
        savings:
          $ref: "#/components/schemas/Savings"
        total_km:
          type: number
        matrix_source:
          type: string
          enum: [graphhopper, osrm, haversine, mixed]
          description: Distance matrix source used for optimization (mixed = combination of sources per cluster)
        geocoder_used:
          type: string
          enum: [yandex, nominatim]
          description: Geocoding service used for address resolution
      required:
        - routes
        - savings
        - total_km
    VehicleRoute:
      type: object
      properties:
        vehicle_name:
          type: string
        stores:
          type: array
          items:
            $ref: "#/components/schemas/RouteStop"
        total_km:
          type: number
        estimated_minutes:
          type: integer
        yandex_url:
          type: string
          description: "Первая ссылка Яндекс.Навигатора (первые 20 точек). Для совместимости."
        yandex_urls:
          type: array
          items:
            type: string
          description: "Список ссылок Яндекс.Навигатора — по одной на каждые 20 точек маршрута."
        whatsapp_url:
          type: string
      required:
        - vehicle_name
        - stores
        - total_km
        - estimated_minutes
        - yandex_url
        - yandex_urls
        - whatsapp_url
    RouteStop:
      type: object
      properties:
        order:
          type: integer
        store_id:
          type: integer
        store_name:
          type: string
        address:
          type: string
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
        arrive_by:
          type: ["string", "null"]
      required:
        - order
        - store_id
        - store_name
        - address
    Savings:
      type: object
      properties:
        optimized_km:
          type: number
        unoptimized_km:
          type: number
        saved_km:
          type: number
        saved_pct:
          type: integer
        saved_fuel_l:
          type: number
        saved_fuel_cost_rub:
          type: integer
          description: "Экономия только на топливе, руб. (saved_fuel_l × цена дизеля)"
        saved_rub_day:
          type: integer
          description: "Полная экономия в день, руб. (топливо + водитель + обслуживание, 31 руб/км × ROAD_FACTOR=1.4)"
        saved_rub_month:
          type: integer
          description: "Полная экономия в месяц = saved_rub_day × 30"
      required:
        - optimized_km
        - unoptimized_km
        - saved_km
        - saved_pct
        - saved_fuel_l
        - saved_fuel_cost_rub
        - saved_rub_day
        - saved_rub_month
...
    RouteSessionItem:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        num_vehicles:
          type: integer
        total_km:
          type: number
        saved_km:
          type: number
        saved_rub:
          type: integer
        num_points:
          type: integer
        created_at:
          type: ["string", "null"]
      required:
        - id
        - date
        - num_vehicles
        - total_km
        - saved_km
        - saved_rub
        - num_points
    RouteSessionList:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/RouteSessionItem"
      required:
        - total
        - page
        - page_size
        - items
...