# lib/api-zod/src/generated/types/routeStop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartRoute API - Delivery route optimization
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStop {
  order: number;
  store_id: number;
  store_name: string;
  address: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  /** @nullable */
  arrive_by?: string | null;
}